Add basic unit tests for filename parsing and WAV generation; update README.md
